room-service/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0 # Added for token handling
python-dotenv==1.0.1
PyJWT==2.8.0 # Added for token handling
psycopg2-binary==2.9.9
requests==2.31.0 # Added for potential inter-service calls
